--- runlab/src/main/java/RunLab/Responces/Responce.java
@@ -1,15 +1,0 @@
-package RunLab.Responces;
-
-public abstract class Responce{
-    protected int status = 400;
-
-    protected String message = "";
-
-    public int getStatus(){
-        return status;
-    }
-
-    public String getMessage(){
-        return message;
-    }
-}
